requirements.txt
python-telegram-bot==21.9
aiosqlite==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1
